Revert "Stable Version Merge"
