--- Source/Assets/AsteroidGenPlugin.meta
@@ -1,5 +1,0 @@
-fileFormatVersion: 2
-guid: d37fe3633ebb40149a4b29c550ee0ec0
-folderAsset: yes
-DefaultImporter:
-  userData:
