--- Source/Assets/AsteroidGenPlugin/Scripts.meta
@@ -1,5 +1,0 @@
-fileFormatVersion: 2
-guid: 1d038ff712514b24d8396fa42ec3bed8
-folderAsset: yes
-DefaultImporter:
-  userData:
