--- Source/Assets/AsteroidGenPlugin/Scripts/AsteroidAnimation.cs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2e334c808f44cc440a3f4977e01563d1
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
