--- Source/Assets/AsteroidGenPlugin/Scripts/AsteroidGeneratorEditior.cs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: bad2aff4711daa44a8dcac7699e32ae0
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
