--- Source/Assets/AsteroidGenPlugin/Scripts/CollisionRaycastScript.cs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 704d31d10b243c443b2a9a39416f18c4
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
