--- Source/Assets/AsteroidGenPlugin/Textures.meta
@@ -1,5 +1,0 @@
-fileFormatVersion: 2
-guid: 0b7db0500d3175c42974ebb8172f9603
-folderAsset: yes
-DefaultImporter:
-  userData:
