+++ Source/ProjectSettings/TagManager.asset
@@ -12,7 +12,7 @@ TagManager:
   Builtin Layer 5: UI
   Builtin Layer 6: 
   Builtin Layer 7: 
-  User Layer 8: Gizmo
+  User Layer 8: 
   User Layer 9: 
   User Layer 10: 
   User Layer 11:
